Wizard: Deduplicate `ManageRepositoriesButton`
This moves the `ManageRepositoriesButton` component to its own file and deduplicates it in the code base.

Tooltip for Upload repositories was also fixed and is now readable again.
